back/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.19
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-modeltranslation==0.19.12
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
future==1.0.0
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.1.0
PyJWT==2.9.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
serializers==0.2.4
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
